+++ java-client-sdk/lib/src/main/java/industries/dot/brease/SDKConfiguration.java
@@ -13,7 +13,8 @@ class SDKConfiguration {
 	public String serverUrl;
 	public int serverIdx = 0;
 	public String language = "java";
-	public String sdkVersion = "0.2.3";
-	public String genVersion = "2.35.9";
+	public String openapiDocVersion = "0.1.0";
+	public String sdkVersion = "0.3.0";
+	public String genVersion = "2.37.0";
 	
 }